feat: add new `gix cat` command.
It only prints things without fuzz.

Inspired by https://youtu.be/JYH5ILv5g1g?si=bHLBPFJiZyRUTl6u&t=211.
